Move pyarrow to an in method import.
This allows client to exclude the pyarrow dep if they don't need it. Saved ~80MB and more compatible with older systems.

Will still get a runtime error if they exclude it, then try to use it.

Still works as expected unless users go out of their way to manually exclude this dependency (I'm not removing the dep, you need to manually exclude it).